chore: test suite clean up
