+++ HarborTests/BackendRecoveryServiceTests.swift
@@ -33,54 +33,6 @@ extension HarborModelAndSafetyTests {
         XCTAssertFalse(fileManager.fileExists(atPath: retainedFolder.path))
     }
 
-    func testAriaPerTorrentOptionsUseAuthoritativeItemLimits() {
-        let settings = DownloadTransferSettings(
-            maxConcurrentDownloads: 3,
-            globalSpeedLimitBytesPerSecond: 900_000,
-            perDownloadSpeedLimitBytesPerSecond: 500_000,
-            globalUploadSpeedLimitBytesPerSecond: 300_000,
-            perDownloadUploadSpeedLimitBytesPerSecond: 200_000,
-            perDownloadConnectionCount: 6
-        )
-        let options = Aria2TorrentService.perDownloadOptions(
-            settings,
-            transferOptions: TorrentTransferOptions(
-                downloadLimitBytesPerSecond: nil,
-                uploadLimitBytesPerSecond: 75_000,
-                shouldSeed: true
-            )
-        )
-
-        XCTAssertEqual(options["max-download-limit"], "0")
-        XCTAssertEqual(options["max-upload-limit"], "75000")
-        XCTAssertEqual(options["max-connection-per-server"], "6")
-        XCTAssertEqual(options["seed-ratio"], "0.0")
-        XCTAssertNil(options["seed-time"])
-
-        let ratioLimitedOptions = Aria2TorrentService.perDownloadOptions(
-            settings,
-            transferOptions: TorrentTransferOptions(
-                downloadLimitBytesPerSecond: nil,
-                uploadLimitBytesPerSecond: nil,
-                shouldSeed: true,
-                seedRatioLimit: 2
-            )
-        )
-        XCTAssertEqual(ratioLimitedOptions["seed-ratio"], "2.0")
-    }
-
-    func testAriaTorrentOptionsDoNotInjectIrrelevantHTTPRecoverySettings() {
-        let options = Aria2TorrentService.perDownloadOptions(
-            .default,
-            transferOptions: nil
-        )
-
-        XCTAssertNil(options["continue"])
-        XCTAssertNil(options["max-tries"])
-        XCTAssertNil(options["retry-wait"])
-        XCTAssertNil(options["bt-stop-timeout"])
-    }
-
     func testOnlyExplicitAriaRPCRejectionMakesMutationFailureDefinitive() {
         XCTAssertFalse(
             Aria2TorrentService.mutationFailureWasExplicitlyRejected(
@@ -404,60 +356,6 @@ extension HarborModelAndSafetyTests {
         )
     }
 
-    func testTorrentSubmissionReservationIsStableUntilExactAcknowledgement() throws {
-        let fileManager = FileManager.default
-        let root = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborTorrentReservation-\(UUID().uuidString)", isDirectory: true)
-        defer { try? fileManager.removeItem(at: root) }
-        let store = TorrentSubmissionReservationStore(
-            fileManager: fileManager,
-            directoryURL: root
-        )
-        let ownerID = UUID()
-        let sourceURL = try XCTUnwrap(
-            URL(string: "magnet:?xt=urn:btih:0123456789abcdef0123456789abcdef01234567")
-        )
-
-        let first = try store.reserve(
-            ownerDownloadID: ownerID,
-            sourceURL: sourceURL,
-            destinationFolderPath: "/tmp/downloads",
-            sourceKind: .magnetLink,
-            shouldSeedAfterDownload: true
-        )
-        let replay = try store.reserve(
-            ownerDownloadID: ownerID,
-            sourceURL: sourceURL,
-            destinationFolderPath: "/tmp/downloads",
-            sourceKind: .magnetLink,
-            shouldSeedAfterDownload: true
-        )
-
-        XCTAssertEqual(first.gid, replay.gid)
-        XCTAssertEqual(first.sourceKind, .magnetLink)
-        XCTAssertEqual(first.shouldSeedAfterDownload, true)
-        XCTAssertEqual(try store.reservations().map(\.gid), [first.gid])
-        XCTAssertNotNil(
-            first.gid.range(of: "^[0-9a-f]{16}$", options: .regularExpression)
-        )
-        XCTAssertThrowsError(
-            try store.reserve(
-                ownerDownloadID: ownerID,
-                sourceURL: sourceURL,
-                destinationFolderPath: "/tmp/other",
-                sourceKind: .magnetLink,
-                shouldSeedAfterDownload: true
-            )
-        )
-        XCTAssertThrowsError(
-            try store.acknowledge(ownerDownloadID: ownerID, gid: "0000000000000000")
-        )
-        XCTAssertEqual(try store.reservation(ownerDownloadID: ownerID)?.gid, first.gid)
-
-        try store.acknowledge(ownerDownloadID: ownerID, gid: first.gid)
-        XCTAssertNil(try store.reservation(ownerDownloadID: ownerID))
-    }
-
     func testManagedChildProcessDrainsFinalOutputBeforeTermination() async throws {
         let state = ManagedProcessTestState()
         let terminated = expectation(description: "child process terminated")
@@ -486,236 +384,4 @@ extension HarborModelAndSafetyTests {
         withExtendedLifetime(process) {}
     }
 
-    func testManagedChildProcessCanAlwaysTerminateAfterSuccessfulConstruction() async throws {
-        let state = ManagedProcessTestState()
-        let terminated = expectation(description: "managed child was terminated")
-        let process = try ManagedChildProcess(
-            executableURL: URL(fileURLWithPath: "/bin/sleep"),
-            arguments: ["30"],
-            environment: ProcessInfo.processInfo.environment,
-            onStdout: { state.appendStdout($0) },
-            onStderr: { state.appendStderr($0) },
-            onTermination: { termination in
-                state.recordTermination(termination)
-                terminated.fulfill()
-            }
-        )
-
-        XCTAssertEqual(getpgid(process.processIdentifier), process.processIdentifier)
-        process.terminate(grace: 0.1)
-        await fulfillment(of: [terminated], timeout: 3)
-        XCTAssertFalse(process.isRunning)
-        XCTAssertFalse(state.snapshot.termination?.isSuccess ?? true)
-    }
-
-
-    func testMediaDownloadArgumentsKeepAutomaticAndExactFormatPathsSeparate() throws {
-        let runtime = MediaRuntimeResolution(
-            ytDlpURL: URL(fileURLWithPath: "/tmp/yt-dlp"),
-            ffmpegURL: URL(fileURLWithPath: "/tmp/ffmpeg"),
-            ffprobeURL: URL(fileURLWithPath: "/tmp/ffprobe")
-        )
-        let sourceURL = try XCTUnwrap(URL(string: "https://example.com/video"))
-        let destinationURL = URL(fileURLWithPath: "/tmp/downloads", isDirectory: true)
-        let temporaryURL = URL(fileURLWithPath: "/tmp/media", isDirectory: true)
-        let completionReceiptURL = URL(
-            fileURLWithPath: "/tmp/media%owned/final-paths.jsonl"
-        )
-
-        let limitedArguments = try MediaDownloadService.downloadArguments(
-            runtime: runtime,
-            sourceURL: sourceURL,
-            destinationFolder: destinationURL,
-            temporaryFolder: temporaryURL,
-            metadata: nil,
-            formatPreference: .bestAvailable,
-            completionReceiptURL: completionReceiptURL,
-            speedLimitBytesPerSecond: 345_678
-        )
-        let unlimitedArguments = try MediaDownloadService.downloadArguments(
-            runtime: runtime,
-            sourceURL: sourceURL,
-            destinationFolder: destinationURL,
-            temporaryFolder: temporaryURL,
-            metadata: nil,
-            formatPreference: .bestAvailable,
-            completionReceiptURL: completionReceiptURL,
-            speedLimitBytesPerSecond: nil
-        )
-        let outputConflictIdentifier = UUID()
-        let collisionSafeArguments = try MediaDownloadService.downloadArguments(
-            runtime: runtime,
-            sourceURL: sourceURL,
-            destinationFolder: destinationURL,
-            temporaryFolder: temporaryURL,
-            metadata: nil,
-            formatPreference: .bestAvailable,
-            outputConflictIdentifier: outputConflictIdentifier,
-            completionReceiptURL: completionReceiptURL,
-            speedLimitBytesPerSecond: nil
-        )
-
-        let limitIndex = try XCTUnwrap(limitedArguments.firstIndex(of: "--limit-rate"))
-        XCTAssertEqual(limitedArguments[limitedArguments.index(after: limitIndex)], "345678")
-        XCTAssertTrue(limitedArguments.contains("--progress"))
-        XCTAssertTrue(unlimitedArguments.contains("--progress"))
-        XCTAssertFalse(unlimitedArguments.contains("--limit-rate"))
-        XCTAssertFalse(limitedArguments.contains("--format"))
-        XCTAssertFalse(unlimitedArguments.contains("--format"))
-        let collisionOutputIndex = try XCTUnwrap(
-            collisionSafeArguments.firstIndex(of: "--output")
-        )
-        XCTAssertEqual(
-            collisionSafeArguments[collisionSafeArguments.index(after: collisionOutputIndex)],
-            "%(title).180B [%(id)s] [Harbor \(outputConflictIdentifier.uuidString)].%(ext)s"
-        )
-        let concurrentDownloadID = UUID()
-        XCTAssertNil(
-            MediaDownloadService.outputIdentifier(
-                requested: nil,
-                downloadID: concurrentDownloadID,
-                hasCompetingDestination: false
-            )
-        )
-        XCTAssertEqual(
-            MediaDownloadService.outputIdentifier(
-                requested: nil,
-                downloadID: concurrentDownloadID,
-                hasCompetingDestination: true
-            ),
-            concurrentDownloadID
-        )
-        XCTAssertEqual(
-            MediaDownloadService.outputIdentifier(
-                requested: outputConflictIdentifier,
-                downloadID: concurrentDownloadID,
-                hasCompetingDestination: true
-            ),
-            outputConflictIdentifier
-        )
-
-        let retryValues = zip(limitedArguments, limitedArguments.dropFirst())
-            .filter { $0.0 == "--retry-sleep" }
-            .map(\.1)
-        XCTAssertEqual(
-            retryValues,
-            [
-                "http:exp=2:60",
-                "fragment:exp=2:60",
-                "file_access:exp=2:60",
-                "extractor:exp=2:60"
-            ]
-        )
-        func value(after option: String) -> String? {
-            guard let optionIndex = limitedArguments.firstIndex(of: option) else {
-                return nil
-            }
-
-            let valueIndex = limitedArguments.index(after: optionIndex)
-            return limitedArguments.indices.contains(valueIndex)
-                ? limitedArguments[valueIndex]
-                : nil
-        }
-        XCTAssertEqual(value(after: "--retries"), "10")
-        XCTAssertEqual(value(after: "--fragment-retries"), "10")
-        XCTAssertEqual(value(after: "--file-access-retries"), "3")
-        XCTAssertEqual(value(after: "--extractor-retries"), "3")
-        let printToFileIndex = try XCTUnwrap(
-            limitedArguments.firstIndex(of: "--print-to-file")
-        )
-        XCTAssertFalse(limitedArguments.contains("--print"))
-        XCTAssertEqual(
-            limitedArguments[limitedArguments.index(after: printToFileIndex)],
-            "after_move:harbor-file:%(filepath)j"
-        )
-        let receiptPathIndex = limitedArguments.index(
-            printToFileIndex,
-            offsetBy: 2
-        )
-        XCTAssertEqual(
-            limitedArguments[receiptPathIndex],
-            "/tmp/media%%owned/final-paths.jsonl"
-        )
-
-        let videoFormat = MediaDownloadFormatOption(
-            formatID: "137",
-            container: "mp4",
-            videoCodec: "avc1.640028",
-            audioCodec: nil,
-            width: 1_920,
-            height: 1_080,
-            framesPerSecond: 30,
-            dynamicRange: "SDR",
-            bitrateKbps: 4_500,
-            estimatedBytes: 9_000_000
-        )
-        let audioFormat = MediaDownloadFormatOption(
-            formatID: "140",
-            container: "m4a",
-            videoCodec: nil,
-            audioCodec: "mp4a.40.2",
-            width: nil,
-            height: nil,
-            framesPerSecond: nil,
-            dynamicRange: nil,
-            bitrateKbps: 128,
-            estimatedBytes: 1_000_000,
-            language: "en",
-            formatNote: "English (Original)",
-            audioChannels: 2,
-            languagePreference: 10
-        )
-        let metadata = MediaDownloadMetadata(
-            title: "Video",
-            platform: "Test",
-            extractorKey: "Test",
-            thumbnailURL: nil,
-            webpageURL: sourceURL,
-            expectedBytes: 10_000_000,
-            mediaType: .video,
-            entryCount: 1,
-            capabilities: MediaDownloadCapabilities(
-                formatOptions: [videoFormat, audioFormat]
-            )
-        )
-        let selection = MediaDownloadFormatSelection(
-            format: videoFormat,
-            audioFormat: audioFormat
-        )
-        let selectedArguments = try MediaDownloadService.downloadArguments(
-            runtime: runtime,
-            sourceURL: sourceURL,
-            destinationFolder: destinationURL,
-            temporaryFolder: temporaryURL,
-            metadata: metadata.persistenceSnapshot,
-            formatPreference: .specific(selection),
-            completionReceiptURL: temporaryURL.appendingPathComponent("final-paths.jsonl"),
-            speedLimitBytesPerSecond: 345_678
-        )
-
-        let formatIndex = try XCTUnwrap(selectedArguments.firstIndex(of: "--format"))
-        XCTAssertEqual(selectedArguments[selectedArguments.index(after: formatIndex)], "137+140")
-        let mergeIndex = try XCTUnwrap(selectedArguments.firstIndex(of: "--merge-output-format"))
-        XCTAssertEqual(selectedArguments[selectedArguments.index(after: mergeIndex)], "mp4")
-        let selectedLimitIndex = try XCTUnwrap(selectedArguments.firstIndex(of: "--limit-rate"))
-        XCTAssertEqual(
-            selectedArguments[selectedArguments.index(after: selectedLimitIndex)],
-            "345678"
-        )
-
-        XCTAssertEqual(
-            selection.displaySummary,
-            "1080p • MP4 • AVC1 + English (Original)"
-        )
-        XCTAssertEqual(
-            MediaDownloadFormatPreference.specific(selection)
-                .initialExpectedBytes(metadataEstimate: 243_768_398),
-            selection.estimatedBytes
-        )
-        XCTAssertEqual(
-            MediaDownloadFormatPreference.bestAvailable
-                .initialExpectedBytes(metadataEstimate: 243_768_398),
-            243_768_398
-        )
-    }
 }
