+++ HarborTests/CompletedHandoffPublicationTests.swift
@@ -194,93 +194,6 @@ extension HarborModelAndSafetyTests {
         await restoredCenter.shutdownForTermination()
     }
 
-    func testValidStagingHandoffReplacesCorruptReadyPackage() throws {
-        let fileManager = FileManager.default
-        let testRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborStagingPromotionTests-\(UUID().uuidString)", isDirectory: true)
-        let handoffRoot = testRoot.appendingPathComponent("Handoffs", isDirectory: true)
-        let payloadURL = testRoot.appendingPathComponent("payload.bin")
-        defer { try? fileManager.removeItem(at: testRoot) }
-        try fileManager.createDirectory(at: testRoot, withIntermediateDirectories: true)
-        try Data("payload".utf8).write(to: payloadURL)
-
-        let store = CompletedDownloadHandoffStore(directoryURL: handoffRoot)
-        let handoff = try store.publish(
-            payloadAt: payloadURL,
-            manifest: CompletedDownloadHandoffManifest(
-                downloadID: UUID(),
-                attemptIdentifier: UUID(),
-                owner: .direct,
-                sourceURL: try XCTUnwrap(URL(string: "https://example.test/payload.bin")),
-                statusCode: 200,
-                mimeType: "application/octet-stream",
-                suggestedFilename: "payload.bin",
-                actualBytes: 7,
-                expectedBytes: 7
-            )
-        )
-        let readyURL = handoff.packageURL
-        let stagingURL = readyURL.deletingPathExtension().appendingPathExtension("staging")
-        try fileManager.moveItem(at: readyURL, to: stagingURL)
-        try fileManager.createDirectory(at: readyURL, withIntermediateDirectories: false)
-        try Data("corrupt".utf8).write(
-            to: readyURL.appendingPathComponent("manifest.json")
-        )
-
-        let entries = try store.entries()
-        XCTAssertEqual(entries.count, 1)
-        guard case let .valid(promoted) = try XCTUnwrap(entries.first) else {
-            return XCTFail("Expected the valid staging package to be promoted")
-        }
-        XCTAssertEqual(try Data(contentsOf: try XCTUnwrap(promoted.payloadURL)), Data("payload".utf8))
-        XCTAssertFalse(fileManager.fileExists(atPath: stagingURL.path))
-    }
-
-    func testFinalizePromotesReadyManifestLeftInStagingDirectory() throws {
-        let fileManager = FileManager.default
-        let testRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborReadyStagingFinalizeTests-\(UUID().uuidString)", isDirectory: true)
-        let handoffRoot = testRoot.appendingPathComponent("Handoffs", isDirectory: true)
-        let payloadURL = testRoot.appendingPathComponent("payload.bin")
-        defer { try? fileManager.removeItem(at: testRoot) }
-        try fileManager.createDirectory(at: testRoot, withIntermediateDirectories: true)
-        try Data("payload".utf8).write(to: payloadURL)
-
-        let store = CompletedDownloadHandoffStore(directoryURL: handoffRoot)
-        let ready = try store.publish(
-            payloadAt: payloadURL,
-            manifest: CompletedDownloadHandoffManifest(
-                downloadID: UUID(),
-                attemptIdentifier: UUID(),
-                owner: .direct,
-                sourceURL: try XCTUnwrap(URL(string: "https://example.test/payload.bin")),
-                statusCode: 200,
-                mimeType: "application/octet-stream",
-                suggestedFilename: "payload.bin",
-                actualBytes: 7,
-                expectedBytes: 7
-            )
-        )
-        let stagingURL = ready.packageURL
-            .deletingPathExtension()
-            .appendingPathExtension("staging")
-        try fileManager.moveItem(at: ready.packageURL, to: stagingURL)
-        let repeatedClaim = CompletedDownloadHandoffClaim(
-            packageURL: stagingURL,
-            manifest: ready.manifest
-        )
-
-        let finalized = try store.finalize(repeatedClaim)
-
-        XCTAssertEqual(finalized.manifest, ready.manifest)
-        XCTAssertEqual(
-            try Data(contentsOf: try XCTUnwrap(finalized.payloadURL)),
-            Data("payload".utf8)
-        )
-        XCTAssertTrue(fileManager.fileExists(atPath: ready.packageURL.path))
-        XCTAssertFalse(fileManager.fileExists(atPath: stagingURL.path))
-    }
-
     func testClaimingHandoffSurvivesReadFailureAndFinishesRecovery() throws {
         let fileManager = FileManager.default
         let testRoot = fileManager.temporaryDirectory
@@ -438,59 +351,4 @@ extension HarborModelAndSafetyTests {
         }
     }
 
-    func testRecordedDestinationCollisionDoesNotCorruptIntactHandoffPayload() throws {
-        let fileManager = FileManager.default
-        let testRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborRecordedDestinationCollision-\(UUID().uuidString)", isDirectory: true)
-        let handoffRoot = testRoot.appendingPathComponent("Handoffs", isDirectory: true)
-        let destinationRoot = testRoot.appendingPathComponent("Destination", isDirectory: true)
-        let payloadURL = testRoot.appendingPathComponent("payload.bin")
-        defer { try? fileManager.removeItem(at: testRoot) }
-        try fileManager.createDirectory(at: destinationRoot, withIntermediateDirectories: true)
-        let payload = Data("completed-payload".utf8)
-        try payload.write(to: payloadURL)
-
-        let store = CompletedDownloadHandoffStore(
-            fileManager: fileManager,
-            directoryURL: handoffRoot
-        )
-        let downloadID = UUID()
-        let attemptIdentifier = UUID()
-        let handoff = try store.publish(
-            payloadAt: payloadURL,
-            manifest: CompletedDownloadHandoffManifest(
-                downloadID: downloadID,
-                attemptIdentifier: attemptIdentifier,
-                owner: .direct,
-                sourceURL: try XCTUnwrap(URL(string: "https://example.test/payload.bin")),
-                statusCode: 200,
-                mimeType: "application/octet-stream",
-                suggestedFilename: "payload.bin",
-                actualBytes: Int64(payload.count),
-                expectedBytes: Int64(payload.count)
-            )
-        )
-        let occupiedURL = destinationRoot.appendingPathComponent("payload.bin")
-        _ = try store.recordDestination(occupiedURL, for: handoff)
-        try fileManager.createSymbolicLink(
-            at: occupiedURL,
-            withDestinationURL: destinationRoot.appendingPathComponent("missing.bin")
-        )
-
-        let recovered = try XCTUnwrap(
-            store.handoff(
-                downloadID: downloadID,
-                attemptIdentifier: attemptIdentifier
-            )
-        )
-        XCTAssertNil(recovered.destinationURL)
-        XCTAssertEqual(try Data(contentsOf: try XCTUnwrap(recovered.payloadURL)), payload)
-
-        let replacementURL = destinationRoot.appendingPathComponent("payload 2.bin")
-        let rerecorded = try store.recordDestination(replacementURL, for: recovered)
-        XCTAssertEqual(rerecorded.manifest.destinationPath, replacementURL.path)
-        XCTAssertNoThrow(
-            try fileManager.destinationOfSymbolicLink(atPath: occupiedURL.path)
-        )
-    }
 }
