+++ HarborTests/DirectDownloadRecoveryIntegrationTests.swift
@@ -129,128 +129,6 @@ extension HarborModelAndSafetyTests {
         )
     }
 
-    func testMissingResumeValidatorRestartsWithoutCombiningRepresentations() async throws {
-        InterruptingRangeURLProtocol.reset(omitValidatorOnFirstResume: true)
-
-        let fileManager = FileManager.default
-        let testRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborMissingResumeValidatorTests-\(UUID().uuidString)", isDirectory: true)
-        let recoveryRoot = testRoot.appendingPathComponent("Recovery", isDirectory: true)
-        defer { try? fileManager.removeItem(at: testRoot) }
-
-        let configuration = URLSessionConfiguration.ephemeral
-        configuration.protocolClasses = [InterruptingRangeURLProtocol.self]
-        let initialFailure = expectation(description: "The initial representation is interrupted")
-        let missingValidatorFailure = expectation(description: "The unverifiable range is rejected")
-        let completion = expectation(description: "A fresh request downloads the replacement representation")
-        let eventState = DirectDownloadTestEventState()
-        let coordinator = DownloadCoordinator(
-            eventHandler: { event in
-                switch event {
-                case let .failed(_, _, failure):
-                    switch eventState.recordFailure(failure) {
-                    case 1: initialFailure.fulfill()
-                    case 2: missingValidatorFailure.fulfill()
-                    default: break
-                    }
-                case let .finished(_, _, handoff):
-                    eventState.recordCompletion(handoff)
-                    completion.fulfill()
-                default:
-                    break
-                }
-            },
-            fileManager: fileManager,
-            recoveryDirectoryURL: recoveryRoot,
-            sessionConfiguration: configuration
-        )
-
-        let id = UUID()
-        let sourceURL = try XCTUnwrap(URL(string: "https://recovery.example.test/file.bin"))
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [initialFailure], timeout: 2)
-
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [missingValidatorFailure], timeout: 2)
-
-        XCTAssertNil(eventState.completedURL)
-        XCTAssertTrue(eventState.failures.last?.requiresFreshStart ?? false)
-        XCTAssertEqual(eventState.failures.last?.recoverableBytes, 0)
-        XCTAssertNil(coordinator.recoverySnapshot(id: id, sourceURL: sourceURL))
-
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [completion], timeout: 2)
-
-        let completedURL = try XCTUnwrap(eventState.completedURL)
-        defer { try? fileManager.removeItem(at: completedURL) }
-        XCTAssertEqual(try Data(contentsOf: completedURL), Data("klmnopqrst".utf8))
-        XCTAssertEqual(
-            InterruptingRangeURLProtocol.capturedHeaders.map(\.range),
-            [nil, "bytes=5-", nil]
-        )
-    }
-
-    func testChangedResumeTotalRestartsWithoutCombiningRepresentations() async throws {
-        InterruptingRangeURLProtocol.reset(changeTotalOnFirstResume: true)
-
-        let fileManager = FileManager.default
-        let testRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborChangedResumeTotalTests-\(UUID().uuidString)", isDirectory: true)
-        let recoveryRoot = testRoot.appendingPathComponent("Recovery", isDirectory: true)
-        defer { try? fileManager.removeItem(at: testRoot) }
-
-        let configuration = URLSessionConfiguration.ephemeral
-        configuration.protocolClasses = [InterruptingRangeURLProtocol.self]
-        let initialFailure = expectation(description: "The initial ten-byte representation is interrupted")
-        let changedTotalFailure = expectation(description: "The contradictory eight-byte total is rejected")
-        let completion = expectation(description: "A fresh request downloads the eight-byte replacement")
-        let eventState = DirectDownloadTestEventState()
-        let coordinator = DownloadCoordinator(
-            eventHandler: { event in
-                switch event {
-                case let .failed(_, _, failure):
-                    switch eventState.recordFailure(failure) {
-                    case 1: initialFailure.fulfill()
-                    case 2: changedTotalFailure.fulfill()
-                    default: break
-                    }
-                case let .finished(_, _, handoff):
-                    eventState.recordCompletion(handoff)
-                    completion.fulfill()
-                default:
-                    break
-                }
-            },
-            fileManager: fileManager,
-            recoveryDirectoryURL: recoveryRoot,
-            sessionConfiguration: configuration
-        )
-
-        let id = UUID()
-        let sourceURL = try XCTUnwrap(URL(string: "https://recovery.example.test/file.bin"))
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [initialFailure], timeout: 2)
-
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [changedTotalFailure], timeout: 2)
-
-        XCTAssertNil(eventState.completedURL)
-        XCTAssertTrue(eventState.failures.last?.requiresFreshStart ?? false)
-        XCTAssertEqual(eventState.failures.last?.recoverableBytes, 0)
-        XCTAssertNil(coordinator.recoverySnapshot(id: id, sourceURL: sourceURL))
-
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [completion], timeout: 2)
-
-        let completedURL = try XCTUnwrap(eventState.completedURL)
-        defer { try? fileManager.removeItem(at: completedURL) }
-        XCTAssertEqual(try Data(contentsOf: completedURL), Data("uvwxyz12".utf8))
-        XCTAssertEqual(
-            InterruptingRangeURLProtocol.capturedHeaders.map(\.range),
-            [nil, "bytes=5-", nil]
-        )
-    }
-
     func testFullResponseFallbackDoesNotReusePartialRepresentationMetadata() async throws {
         InterruptingRangeURLProtocol.reset(fallbackToFullResponseOnFirstResume: true)
 
@@ -301,67 +179,6 @@ extension HarborModelAndSafetyTests {
         XCTAssertEqual(
             InterruptingRangeURLProtocol.capturedHeaders.map(\.range),
             [nil, "bytes=5-"]
-        )
-    }
-
-    func testContradictoryUnsatisfiedRangeCannotCertifyIncompletePartial() async throws {
-        InterruptingRangeURLProtocol.reset(contradictSavedTotalOnFirstResume: true)
-
-        let fileManager = FileManager.default
-        let testRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborContradictory416Tests-\(UUID().uuidString)", isDirectory: true)
-        let recoveryRoot = testRoot.appendingPathComponent("Recovery", isDirectory: true)
-        defer { try? fileManager.removeItem(at: testRoot) }
-
-        let configuration = URLSessionConfiguration.ephemeral
-        configuration.protocolClasses = [InterruptingRangeURLProtocol.self]
-        let initialFailure = expectation(description: "The initial ten-byte representation is interrupted")
-        let contradictoryRangeFailure = expectation(description: "The contradictory five-byte total is rejected")
-        let completion = expectation(description: "A fresh request downloads the five-byte replacement")
-        let eventState = DirectDownloadTestEventState()
-        let coordinator = DownloadCoordinator(
-            eventHandler: { event in
-                switch event {
-                case let .failed(_, _, failure):
-                    switch eventState.recordFailure(failure) {
-                    case 1: initialFailure.fulfill()
-                    case 2: contradictoryRangeFailure.fulfill()
-                    default: break
-                    }
-                case let .finished(_, _, handoff):
-                    eventState.recordCompletion(handoff)
-                    completion.fulfill()
-                default:
-                    break
-                }
-            },
-            fileManager: fileManager,
-            recoveryDirectoryURL: recoveryRoot,
-            sessionConfiguration: configuration
-        )
-
-        let id = UUID()
-        let sourceURL = try XCTUnwrap(URL(string: "https://recovery.example.test/file.bin"))
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [initialFailure], timeout: 2)
-
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [contradictoryRangeFailure], timeout: 2)
-
-        XCTAssertNil(eventState.completedURL)
-        XCTAssertTrue(eventState.failures.last?.requiresFreshStart ?? false)
-        XCTAssertEqual(eventState.failures.last?.recoverableBytes, 0)
-        XCTAssertNil(coordinator.recoverySnapshot(id: id, sourceURL: sourceURL))
-
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [completion], timeout: 2)
-
-        let completedURL = try XCTUnwrap(eventState.completedURL)
-        defer { try? fileManager.removeItem(at: completedURL) }
-        XCTAssertEqual(try Data(contentsOf: completedURL), Data("uvwxy".utf8))
-        XCTAssertEqual(
-            InterruptingRangeURLProtocol.capturedHeaders.map(\.range),
-            [nil, "bytes=5-", nil]
         )
     }
 
@@ -483,57 +300,6 @@ extension HarborModelAndSafetyTests {
         XCTAssertEqual(InterruptingRangeURLProtocol.capturedHeaders.map(\.range), ["bytes=5-"])
     }
 
-    func testIncompleteRangeResponseRetainsPartialInsteadOfFinishing() async throws {
-        InterruptingRangeURLProtocol.reset(shortFirstResume: true)
-
-        let fileManager = FileManager.default
-        let testRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborShortRangeTests-\(UUID().uuidString)", isDirectory: true)
-        let recoveryRoot = testRoot.appendingPathComponent("Recovery", isDirectory: true)
-        defer { try? fileManager.removeItem(at: testRoot) }
-
-        let configuration = URLSessionConfiguration.ephemeral
-        configuration.protocolClasses = [InterruptingRangeURLProtocol.self]
-        let firstFailure = expectation(description: "The initial transfer is interrupted")
-        let incompleteRangeFailure = expectation(description: "The capped range is rejected")
-        let eventState = DirectDownloadTestEventState()
-        let coordinator = DownloadCoordinator(
-            eventHandler: { event in
-                switch event {
-                case let .failed(_, _, failure):
-                    switch eventState.recordFailure(failure) {
-                    case 1: firstFailure.fulfill()
-                    case 2: incompleteRangeFailure.fulfill()
-                    default: break
-                    }
-                case let .finished(_, _, handoff):
-                    eventState.recordCompletion(handoff)
-                default:
-                    break
-                }
-            },
-            fileManager: fileManager,
-            recoveryDirectoryURL: recoveryRoot,
-            sessionConfiguration: configuration
-        )
-
-        let id = UUID()
-        let sourceURL = try XCTUnwrap(URL(string: "https://recovery.example.test/file.bin"))
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [firstFailure], timeout: 2)
-
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [incompleteRangeFailure], timeout: 2)
-
-        XCTAssertNil(eventState.completedURL)
-        XCTAssertTrue(eventState.failures.last?.isRetryable ?? false)
-        XCTAssertEqual(eventState.failures.last?.recoverableBytes, 8)
-        XCTAssertEqual(
-            coordinator.recoverySnapshot(id: id, sourceURL: sourceURL)?.bytesWritten,
-            8
-        )
-    }
-
     func testOverlongRangeResponseCannotPublishUndeclaredBytesOnRetry() async throws {
         InterruptingRangeURLProtocol.reset(overlongFirstResume: true)
 
@@ -594,67 +360,6 @@ extension HarborModelAndSafetyTests {
         XCTAssertEqual(requests.count, 3)
         XCTAssertEqual(requests[1].range, "bytes=5-")
         XCTAssertNil(requests[2].range)
-    }
-
-    func testUnknownRangeTotalPreservesSavedLengthAndRejectsTruncation() async throws {
-        InterruptingRangeURLProtocol.reset(unknownTotalFirstResume: true)
-
-        let fileManager = FileManager.default
-        let testRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborUnknownRangeTotalTests-\(UUID().uuidString)", isDirectory: true)
-        let recoveryRoot = testRoot.appendingPathComponent("Recovery", isDirectory: true)
-        defer { try? fileManager.removeItem(at: testRoot) }
-
-        let configuration = URLSessionConfiguration.ephemeral
-        configuration.protocolClasses = [InterruptingRangeURLProtocol.self]
-        let firstFailure = expectation(description: "The initial transfer is interrupted")
-        let incompleteRangeFailure = expectation(description: "The unknown-total range is rejected")
-        let completion = expectation(description: "The fresh retry completes")
-        let eventState = DirectDownloadTestEventState()
-        let coordinator = DownloadCoordinator(
-            eventHandler: { event in
-                switch event {
-                case let .failed(_, _, failure):
-                    switch eventState.recordFailure(failure) {
-                    case 1: firstFailure.fulfill()
-                    case 2: incompleteRangeFailure.fulfill()
-                    default: break
-                    }
-                case let .finished(_, _, handoff):
-                    eventState.recordCompletion(handoff)
-                    completion.fulfill()
-                default:
-                    break
-                }
-            },
-            fileManager: fileManager,
-            recoveryDirectoryURL: recoveryRoot,
-            sessionConfiguration: configuration
-        )
-
-        let id = UUID()
-        let sourceURL = try XCTUnwrap(URL(string: "https://recovery.example.test/file.bin"))
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [firstFailure], timeout: 2)
-
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [incompleteRangeFailure], timeout: 2)
-
-        XCTAssertNil(eventState.completedURL)
-        XCTAssertTrue(eventState.failures.last?.isRetryable ?? false)
-        XCTAssertEqual(eventState.failures.last?.recoverableBytes, 0)
-        XCTAssertNil(coordinator.recoverySnapshot(id: id, sourceURL: sourceURL))
-
-        try coordinator.startDownload(id: id, sourceURL: sourceURL, resumeData: nil)
-        await fulfillment(of: [completion], timeout: 2)
-
-        let completedURL = try XCTUnwrap(eventState.completedURL)
-        defer { try? fileManager.removeItem(at: completedURL) }
-        XCTAssertEqual(try Data(contentsOf: completedURL), Data("abcdefghij".utf8))
-        XCTAssertEqual(
-            InterruptingRangeURLProtocol.capturedHeaders.map(\.range),
-            [nil, "bytes=5-", nil]
-        )
     }
 
     func testValidatorlessInterruptedDownloadRestartsWithoutRange() async throws {
