+++ HarborTests/DownloadLifecycleMutationTests.swift
@@ -117,204 +117,6 @@ extension HarborModelAndSafetyTests {
         await center.shutdownForTermination()
     }
 
-    func testPendingDataRemovalFinishesAfterCrashBetweenTrashAndRecordSave() async throws {
-        let suiteName = "HarborTests.PendingDataRemovalReplay.\(UUID().uuidString)"
-        let userDefaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        userDefaults.removePersistentDomain(forName: suiteName)
-        defer { userDefaults.removePersistentDomain(forName: suiteName) }
-
-        let fileManager = FileManager.default
-        let testRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborPendingDataRemovalReplayTests-\(UUID().uuidString)", isDirectory: true)
-        let persistenceRoot = testRoot.appendingPathComponent("Persistence", isDirectory: true)
-        let pendingRoot = testRoot.appendingPathComponent("PendingRemoval", isDirectory: true)
-        let directRoot = testRoot.appendingPathComponent("DirectRecovery", isDirectory: true)
-        let browserRoot = testRoot.appendingPathComponent("BrowserRecovery", isDirectory: true)
-        let handoffRoot = testRoot.appendingPathComponent("Handoffs", isDirectory: true)
-        let destinationRoot = testRoot.appendingPathComponent("Destination", isDirectory: true)
-        let payloadURL = destinationRoot.appendingPathComponent("archive.bin")
-        defer { try? fileManager.removeItem(at: testRoot) }
-        try fileManager.createDirectory(at: destinationRoot, withIntermediateDirectories: true)
-        try Data("already-trashed".utf8).write(to: payloadURL)
-
-        let item = DownloadItem(
-            sourceURL: try XCTUnwrap(URL(string: "https://example.test/archive.bin")),
-            sourceKind: .directURL,
-            backend: .urlSession,
-            preferredFilename: "archive.bin",
-            destinationFolderPath: destinationRoot.path,
-            fileLocationPath: payloadURL.path,
-            status: .completed,
-            progress: 1,
-            bytesWritten: 15,
-            expectedBytes: 15,
-            finishedAt: .now
-        )
-        let persistence = DownloadPersistence(directoryURL: persistenceRoot)
-        try await persistence.save([item.makeRecord()])
-        let pendingStore = PendingDownloadDataRemovalStore(
-            fileManager: fileManager,
-            directoryURL: pendingRoot
-        )
-        try pendingStore.publish(record: item.makeRecord())
-
-        // This is the crash boundary: the payload move completed, but the old
-        // record was still the last durable download-list snapshot.
-        try fileManager.removeItem(at: payloadURL)
-
-        let settings = AppSettingsStore(userDefaults: userDefaults)
-        settings.startDownloadsAutomatically = false
-        let center = DownloadCenter(
-            settings: settings,
-            persistence: persistence,
-            directRecoveryDirectoryURL: directRoot,
-            completedHandoffDirectoryURL: handoffRoot,
-            browserRecoveryDirectoryURL: browserRoot,
-            pendingDataRemovalDirectoryURL: pendingRoot
-        )
-        await center.initializeIfNeeded()
-
-        XCTAssertTrue(center.downloads.isEmpty)
-        let replayedRecords = try await persistence.load()
-        XCTAssertTrue(replayedRecords.isEmpty)
-        XCTAssertTrue(try pendingStore.entries().isEmpty)
-        XCTAssertFalse(fileManager.fileExists(atPath: payloadURL.path))
-        let didShutDown = await center.shutdownForTermination()
-        XCTAssertTrue(didShutDown)
-    }
-
-    func testInterruptedDataRemovalDoesNotDeleteReplacementAtSamePath() async throws {
-        let suiteName = "HarborTests.PendingDataRemovalReplacement.\(UUID().uuidString)"
-        let userDefaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        userDefaults.removePersistentDomain(forName: suiteName)
-        defer { userDefaults.removePersistentDomain(forName: suiteName) }
-
-        let fileManager = FileManager.default
-        let testRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborPendingRemovalReplacement-\(UUID().uuidString)", isDirectory: true)
-        let persistenceRoot = testRoot.appendingPathComponent("Persistence", isDirectory: true)
-        let pendingRoot = testRoot.appendingPathComponent("PendingRemoval", isDirectory: true)
-        let destinationRoot = testRoot.appendingPathComponent("Destination", isDirectory: true)
-        let payloadURL = destinationRoot.appendingPathComponent("archive.bin")
-        defer { try? fileManager.removeItem(at: testRoot) }
-        try fileManager.createDirectory(at: destinationRoot, withIntermediateDirectories: true)
-        try Data("original-download".utf8).write(to: payloadURL)
-
-        let item = DownloadItem(
-            sourceURL: try XCTUnwrap(URL(string: "https://example.test/archive.bin")),
-            sourceKind: .directURL,
-            backend: .urlSession,
-            preferredFilename: "archive.bin",
-            destinationFolderPath: destinationRoot.path,
-            fileLocationPath: payloadURL.path,
-            status: .completed,
-            progress: 1,
-            bytesWritten: 17,
-            expectedBytes: 17,
-            finishedAt: .now
-        )
-        let persistence = DownloadPersistence(directoryURL: persistenceRoot)
-        try await persistence.save([item.makeRecord()])
-        let pendingStore = PendingDownloadDataRemovalStore(
-            fileManager: fileManager,
-            directoryURL: pendingRoot
-        )
-        try pendingStore.publish(record: item.makeRecord(), removingData: true)
-        _ = try pendingStore.markPayloadDeletionStarted(downloadID: item.id)
-
-        // Model the ambiguous crash boundary: Harbor may have moved the
-        // original away, and another process then created a new file at the
-        // same pathname before Harbor relaunched.
-        try fileManager.removeItem(at: payloadURL)
-        let replacement = Data("replacement-must-survive".utf8)
-        try replacement.write(to: payloadURL)
-
-        let settings = AppSettingsStore(userDefaults: userDefaults)
-        settings.startDownloadsAutomatically = false
-        let center = DownloadCenter(
-            settings: settings,
-            persistence: persistence,
-            directRecoveryDirectoryURL: testRoot.appendingPathComponent("DirectRecovery"),
-            completedHandoffDirectoryURL: testRoot.appendingPathComponent("Handoffs"),
-            browserRecoveryDirectoryURL: testRoot.appendingPathComponent("BrowserRecovery"),
-            pendingDataRemovalDirectoryURL: pendingRoot
-        )
-        await center.initializeIfNeeded()
-
-        let restored = try XCTUnwrap(center.downloads.first { $0.id == item.id })
-        XCTAssertEqual(restored.status, .completed)
-        XCTAssertNotNil(restored.lastError)
-        XCTAssertEqual(try Data(contentsOf: payloadURL), replacement)
-        XCTAssertTrue(try pendingStore.entries().isEmpty)
-        XCTAssertEqual(center.activeAlert?.title, "Download Data Was Left in Place")
-        let persistedIDs = try await persistence.load().map(\.id)
-        XCTAssertEqual(persistedIDs, [item.id])
-        let shutdownSucceeded = await center.shutdownForTermination()
-        XCTAssertTrue(shutdownSucceeded)
-    }
-
-    func testDataRemovalJournalFailureLeavesPayloadAndRecordUntouched() async throws {
-        let suiteName = "HarborTests.PendingDataRemovalPublishFailure.\(UUID().uuidString)"
-        let userDefaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        userDefaults.removePersistentDomain(forName: suiteName)
-        defer { userDefaults.removePersistentDomain(forName: suiteName) }
-
-        let fileManager = FileManager.default
-        let testRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborPendingDataRemovalFailureTests-\(UUID().uuidString)", isDirectory: true)
-        let persistenceRoot = testRoot.appendingPathComponent("Persistence", isDirectory: true)
-        let blockedPendingRoot = testRoot.appendingPathComponent("not-a-directory")
-        let destinationRoot = testRoot.appendingPathComponent("Destination", isDirectory: true)
-        let payloadURL = destinationRoot.appendingPathComponent("archive.bin")
-        defer { try? fileManager.removeItem(at: testRoot) }
-        try fileManager.createDirectory(at: destinationRoot, withIntermediateDirectories: true)
-        try Data("must-remain-owned".utf8).write(to: payloadURL)
-        try Data("occupied".utf8).write(to: blockedPendingRoot)
-
-        let item = DownloadItem(
-            sourceURL: try XCTUnwrap(URL(string: "https://example.test/archive.bin")),
-            sourceKind: .directURL,
-            backend: .urlSession,
-            preferredFilename: "archive.bin",
-            destinationFolderPath: destinationRoot.path,
-            fileLocationPath: payloadURL.path,
-            status: .completed,
-            progress: 1,
-            bytesWritten: 17,
-            expectedBytes: 17,
-            finishedAt: .now
-        )
-        let persistence = DownloadPersistence(directoryURL: persistenceRoot)
-        try await persistence.save([item.makeRecord()])
-        let center = DownloadCenter(
-            settings: AppSettingsStore(userDefaults: userDefaults),
-            persistence: persistence,
-            directRecoveryDirectoryURL: testRoot.appendingPathComponent("DirectRecovery"),
-            completedHandoffDirectoryURL: testRoot.appendingPathComponent("Handoffs"),
-            browserRecoveryDirectoryURL: testRoot.appendingPathComponent("BrowserRecovery"),
-            pendingDataRemovalDirectoryURL: blockedPendingRoot
-        )
-        center.downloads = [item]
-
-        center.removeDownloadsAndData(ids: [item.id])
-        for _ in 0 ..< 200 {
-            if center.activeAlert != nil {
-                break
-            }
-            try await Task.sleep(for: .milliseconds(10))
-        }
-
-        XCTAssertEqual(center.downloads.map(\.id), [item.id])
-        XCTAssertTrue(fileManager.fileExists(atPath: payloadURL.path))
-        XCTAssertEqual(try Data(contentsOf: payloadURL), Data("must-remain-owned".utf8))
-        XCTAssertEqual(center.activeAlert?.title, "Couldn’t Prepare Download Data Removal")
-        let retainedRecords = try await persistence.load()
-        XCTAssertEqual(retainedRecords.map(\.id), [item.id])
-        let didShutDown = await center.shutdownForTermination()
-        XCTAssertTrue(didShutDown)
-    }
-
-
     func testTorrentPauseFailureRestoresStatusAndKeepsQueueSlot() async throws {
         struct ExpectedPauseFailure: LocalizedError {
             var errorDescription: String? { "Expected torrent pause failure" }
@@ -426,113 +228,6 @@ extension HarborModelAndSafetyTests {
         XCTAssertEqual(persistedItem.backendIdentifier, "torrent-gid")
 
         await center.shutdownForTermination()
-    }
-
-    func testTorrentRetryDoesNotClearIdentifierBeforeConfirmedCleanup() async throws {
-        struct ExpectedCleanupFailure: LocalizedError {
-            var errorDescription: String? { "Expected retry cleanup failure" }
-        }
-
-        let suiteName = "HarborTests.TorrentRetryCleanup.\(UUID().uuidString)"
-        let userDefaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        userDefaults.removePersistentDomain(forName: suiteName)
-        defer { userDefaults.removePersistentDomain(forName: suiteName) }
-
-        let fileManager = FileManager.default
-        let persistenceRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborTorrentRetryCleanupTests-\(UUID().uuidString)", isDirectory: true)
-        defer { try? fileManager.removeItem(at: persistenceRoot) }
-        let cleanupGate = AsyncTestGate()
-        let center = DownloadCenter(
-            settings: AppSettingsStore(userDefaults: userDefaults),
-            persistence: DownloadPersistence(directoryURL: persistenceRoot),
-            torrentRemoveOperation: { _, _ in
-                await cleanupGate.wait()
-                throw ExpectedCleanupFailure()
-            }
-        )
-        let item = DownloadItem(
-            sourceURL: try XCTUnwrap(URL(string: "magnet:?xt=urn:btih:0123456789abcdef0123456789abcdef01234567")),
-            sourceKind: .magnetLink,
-            backend: .aria2,
-            preferredFilename: nil,
-            destinationFolderPath: "/tmp",
-            status: .failed
-        )
-        item.backendIdentifier = "old-torrent-gid"
-        center.downloads = [item]
-
-        center.retryDownload(id: item.id)
-        XCTAssertEqual(item.status, .preparing)
-        XCTAssertEqual(item.backendIdentifier, "old-torrent-gid")
-
-        await cleanupGate.release()
-        for _ in 0 ..< 100 {
-            if item.lastError == "Expected retry cleanup failure" {
-                break
-            }
-            try await Task.sleep(for: .milliseconds(10))
-        }
-
-        XCTAssertEqual(item.status, .failed)
-        XCTAssertEqual(item.backendIdentifier, "old-torrent-gid")
-        XCTAssertEqual(center.activeAlert?.message, "Expected retry cleanup failure")
-
-        await center.shutdownForTermination()
-    }
-
-    func testTorrentPauseDuringRetryCleanupSettlesPausedWithoutStartingReplacement() async throws {
-        let suiteName = "HarborTests.TorrentRetryPause.\(UUID().uuidString)"
-        let userDefaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        userDefaults.removePersistentDomain(forName: suiteName)
-        defer { userDefaults.removePersistentDomain(forName: suiteName) }
-
-        let fileManager = FileManager.default
-        let testRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborTorrentRetryPause-\(UUID().uuidString)", isDirectory: true)
-        defer { try? fileManager.removeItem(at: testRoot) }
-        let cleanupEntered = AsyncTestGate()
-        let releaseCleanup = AsyncTestGate()
-        let center = DownloadCenter(
-            settings: AppSettingsStore(userDefaults: userDefaults),
-            persistence: DownloadPersistence(
-                directoryURL: testRoot.appendingPathComponent("Persistence")
-            ),
-            pendingDataRemovalDirectoryURL: testRoot.appendingPathComponent("PendingRemovals"),
-            torrentRemoveOperation: { _, _ in
-                await cleanupEntered.release()
-                await releaseCleanup.wait()
-            }
-        )
-        let item = DownloadItem(
-            sourceURL: try XCTUnwrap(
-                URL(string: "magnet:?xt=urn:btih:0123456789abcdef0123456789abcdef01234567")
-            ),
-            sourceKind: .magnetLink,
-            backend: .aria2,
-            preferredFilename: nil,
-            destinationFolderPath: "/tmp",
-            status: .failed
-        )
-        item.backendIdentifier = "old-torrent-gid"
-        center.downloads = [item]
-
-        center.retryDownload(id: item.id)
-        await cleanupEntered.wait()
-        XCTAssertEqual(item.status, .preparing)
-
-        center.togglePauseResume(id: item.id)
-        await releaseCleanup.release()
-        for _ in 0 ..< 100 where item.status != .paused {
-            try await Task.sleep(for: .milliseconds(10))
-        }
-
-        XCTAssertEqual(item.status, .paused)
-        XCTAssertNil(item.backendIdentifier)
-        XCTAssertEqual(item.bytesWritten, 0)
-        XCTAssertEqual(item.progress, 0)
-        let shutdownSucceeded = await center.shutdownForTermination()
-        XCTAssertTrue(shutdownSucceeded)
     }
 
     func testCancelledDirectRetryDoesNotStartBeforeConfirmedCleanup() async throws {
@@ -693,7 +388,6 @@ extension HarborModelAndSafetyTests {
             .appendingPathComponent("HarborClearFailedSaveFailureTests-\(UUID().uuidString)", isDirectory: true)
         let persistenceRoot = testRoot.appendingPathComponent("Persistence", isDirectory: true)
         let recoveryRoot = testRoot.appendingPathComponent("Recovery", isDirectory: true)
-        let pendingRemovalRoot = testRoot.appendingPathComponent("PendingRemovals", isDirectory: true)
         defer { try? fileManager.removeItem(at: testRoot) }
 
         let sourceURL = try XCTUnwrap(URL(string: "https://example.test/archive.bin"))
@@ -728,7 +422,6 @@ extension HarborModelAndSafetyTests {
             settings: AppSettingsStore(userDefaults: userDefaults),
             persistence: DownloadPersistence(directoryURL: persistenceRoot),
             directRecoveryDirectoryURL: recoveryRoot,
-            pendingDataRemovalDirectoryURL: pendingRemovalRoot,
             recordSaveOperation: { _, _, _ in
                 throw ExpectedSaveFailure()
             }
@@ -745,44 +438,6 @@ extension HarborModelAndSafetyTests {
 
         await center.shutdownForTermination()
     }
-
-    func testClearFailedCannotBeUndoneByCancelledDebouncedSave() async throws {
-        let suiteName = "HarborTests.ClearFailedDebounce.\(UUID().uuidString)"
-        let userDefaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        userDefaults.removePersistentDomain(forName: suiteName)
-        defer { userDefaults.removePersistentDomain(forName: suiteName) }
-
-        let fileManager = FileManager.default
-        let persistenceRoot = fileManager.temporaryDirectory
-            .appendingPathComponent("HarborClearFailedDebounceTests-\(UUID().uuidString)", isDirectory: true)
-        defer { try? fileManager.removeItem(at: persistenceRoot) }
-
-        let failedItem = DownloadItem(
-            sourceURL: try XCTUnwrap(URL(string: "https://example.test/archive.bin")),
-            sourceKind: .directURL,
-            backend: .urlSession,
-            preferredFilename: nil,
-            destinationFolderPath: "/tmp",
-            status: .failed
-        )
-        let persistence = DownloadPersistence(directoryURL: persistenceRoot)
-        let center = DownloadCenter(
-            settings: AppSettingsStore(userDefaults: userDefaults),
-            persistence: persistence
-        )
-        center.downloads = [failedItem]
-
-        center.setDownloadLimitOverride(.unlimited, for: failedItem.id)
-        await center.clearFailed()
-        try await Task.sleep(for: .milliseconds(350))
-
-        XCTAssertTrue(center.downloads.isEmpty)
-        let persistedRecords = try await persistence.load()
-        XCTAssertTrue(persistedRecords.isEmpty)
-
-        await center.shutdownForTermination()
-    }
-
 
     func testShutdownRefusesTerminationWhenTorrentRecoveryCannotBeSaved() async throws {
         struct ExpectedTorrentShutdownFailure: LocalizedError {
