+++ HarborTests/MediaAndModelTests.swift
@@ -147,77 +147,6 @@ extension HarborModelAndSafetyTests {
         )
     }
 
-    func testMediaRecordPersistsSelectionWithoutFormatCatalog() throws {
-        let sourceURL = try XCTUnwrap(URL(string: "https://example.com/video"))
-        let videoFormat = MediaDownloadFormatOption(
-            formatID: "137",
-            container: "mp4",
-            videoCodec: "avc1.640028",
-            audioCodec: nil,
-            width: 1_920,
-            height: 1_080,
-            framesPerSecond: 30,
-            dynamicRange: "SDR",
-            bitrateKbps: 4_500,
-            estimatedBytes: 9_000_000
-        )
-        let audioFormat = MediaDownloadFormatOption(
-            formatID: "140",
-            container: "m4a",
-            videoCodec: nil,
-            audioCodec: "mp4a.40.2",
-            width: nil,
-            height: nil,
-            framesPerSecond: nil,
-            dynamicRange: nil,
-            bitrateKbps: 128,
-            estimatedBytes: 1_000_000,
-            language: "en",
-            formatNote: "English (Original)",
-            audioChannels: 2,
-            languagePreference: 10
-        )
-        let metadata = MediaDownloadMetadata(
-            title: "Video",
-            platform: "Test",
-            extractorKey: "Test",
-            thumbnailURL: nil,
-            webpageURL: sourceURL,
-            expectedBytes: 10_000_000,
-            mediaType: .video,
-            entryCount: 1,
-            capabilities: MediaDownloadCapabilities(
-                formatOptions: [videoFormat, audioFormat]
-            )
-        )
-        let selection = MediaDownloadFormatSelection(
-            format: videoFormat,
-            audioFormat: audioFormat
-        )
-        let item = DownloadItem(
-            sourceURL: sourceURL,
-            sourceKind: .mediaURL,
-            backend: .ytDlp,
-            preferredFilename: nil,
-            destinationFolderPath: "/tmp/downloads",
-            status: .paused,
-            mediaMetadata: metadata,
-            mediaFormatPreference: .specific(selection)
-        )
-
-        let record = item.makeRecord()
-
-        XCTAssertEqual(record.mediaMetadata?.capabilities, .unavailable)
-        XCTAssertEqual(record.mediaFormatPreference, .specific(selection))
-        XCTAssertEqual(
-            item.mediaMetadata?.capabilities.formatOptions,
-            [videoFormat, audioFormat]
-        )
-
-        let encodedRecord = try JSONEncoder().encode(record)
-        XCTAssertFalse(String(decoding: encodedRecord, as: UTF8.self).contains("formatOptions"))
-    }
-
     func testMediaRecoveryBarriersPersistAndDefaultSafelyForLegacyRecords() throws {
         let outputConflictIdentifier = UUID()
         let item = DownloadItem(
@@ -330,122 +259,6 @@ extension HarborModelAndSafetyTests {
         XCTAssertTrue(secondShutdownSucceeded)
     }
 
-    func testUnavailableExactMediaFormatHasDedicatedError() {
-        XCTAssertEqual(
-            MediaDownloadErrorClassifier.message(
-                from: "ERROR: [youtube] Requested format is not available"
-            ),
-            MediaDownloadErrorClassifier.selectedFormatUnavailableMessage
-        )
-    }
-
-    func testQuickLookRequiresCompletedExistingLocalFiles() throws {
-        let suiteName = "HarborTests.QuickLook.\(UUID().uuidString)"
-        let userDefaults = UserDefaults(suiteName: suiteName)!
-        userDefaults.removePersistentDomain(forName: suiteName)
-        defer { userDefaults.removePersistentDomain(forName: suiteName) }
-
-        let fileURL = FileManager.default.temporaryDirectory
-            .appendingPathComponent("Harbor-Quick-Look-\(UUID().uuidString).txt")
-        try Data("Preview".utf8).write(to: fileURL)
-        defer { try? FileManager.default.removeItem(at: fileURL) }
-
-        let previewService = FakeQuickLookPreviewService()
-        let center = DownloadCenter(
-            settings: AppSettingsStore(userDefaults: userDefaults),
-            quickLookPreviewService: previewService
-        )
-        let item = makeTorrentItem(
-            sourceURL: URL(fileURLWithPath: "/tmp/preview.torrent"),
-            sourceKind: .torrentFile,
-            fileLocationPath: fileURL.path
-        )
-        center.downloads = [item]
-        center.selectedDownloadID = item.id
-
-        XCTAssertTrue(center.canQuickLookSelectedDownloads)
-
-        center.quickLookSelectedDownloads()
-
-        XCTAssertEqual(previewService.previewedURLs, [fileURL])
-        XCTAssertEqual(item.status, .completed)
-
-        item.status = .paused
-        XCTAssertFalse(center.canQuickLookSelectedDownloads)
-
-        item.status = .completed
-        try FileManager.default.removeItem(at: fileURL)
-        XCTAssertFalse(center.canQuickLookSelectedDownloads)
-
-        center.quickLookSelectedDownloads()
-        XCTAssertEqual(center.activeAlert?.title, "Quick Look Unavailable")
-    }
-
-    func testLegacyCompletedTorrentDoesNotSeed() throws {
-        let record = try legacyTorrentRecord(status: .completed)
-
-        XCTAssertFalse(record.shouldSeedAfterDownload)
-        XCTAssertFalse(record.removeOriginalTorrentAfterImport)
-        XCTAssertTrue(record.completionNotificationDelivered)
-        XCTAssertEqual(record.downloadLimitOverride, .inherit)
-        XCTAssertEqual(record.uploadLimitOverride, .inherit)
-        XCTAssertTrue(record.torrentPayloadPaths.isEmpty)
-    }
-
-    func testLegacyUnfinishedTorrentSeeds() throws {
-        let record = try legacyTorrentRecord(status: .downloading)
-
-        XCTAssertTrue(record.shouldSeedAfterDownload)
-    }
-
-    func testTorrentDisplayNamePrefersSemanticMetadata() {
-        let item = makeTorrentItem(
-            sourceURL: URL(fileURLWithPath: "/tmp/ubuntu.torrent"),
-            sourceKind: .torrentFile,
-            metadataName: "Ubuntu 26.04",
-            fileLocationPath: "/tmp/download.html"
-        )
-
-        XCTAssertEqual(item.displayName, "Ubuntu 26.04")
-    }
-
-    func testTorrentDisplayNameUsesTorrentFilenameBeforePayloadFilename() {
-        let item = makeTorrentItem(
-            sourceURL: URL(fileURLWithPath: "/tmp/ubuntu-desktop.torrent"),
-            sourceKind: .torrentFile,
-            fileLocationPath: "/tmp/download.html"
-        )
-
-        XCTAssertEqual(item.displayName, "ubuntu-desktop")
-    }
-
-    func testMagnetDisplayNameWinsOverPayloadFilename() throws {
-        let sourceURL = try XCTUnwrap(
-            URL(string: "magnet:?xt=urn:btih:ABC123&dn=Ubuntu%2026.04")
-        )
-        let item = makeTorrentItem(
-            sourceURL: sourceURL,
-            sourceKind: .magnetLink,
-            fileLocationPath: "/tmp/download.html"
-        )
-
-        XCTAssertEqual(item.displayName, "Ubuntu 26.04")
-    }
-
-    func testSeedingStatusIsActiveButDoesNotConsumeDownloadConcurrency() {
-        let item = makeTorrentItem(
-            sourceURL: URL(fileURLWithPath: "/tmp/ubuntu.torrent"),
-            sourceKind: .torrentFile,
-            status: .seeding
-        )
-
-        XCTAssertFalse(DownloadStatus.seeding.isTerminal)
-        XCTAssertFalse(DownloadStatus.seeding.isRunning)
-        XCTAssertTrue(DownloadStatus.allCases.contains(.seeding))
-        XCTAssertTrue(DownloadFilter.active.includes(item))
-        XCTAssertTrue(item.canPause)
-    }
-
     func testBrowserResumeDataPersistsSeparatelyFromURLSessionResumeData() throws {
         let browserResumeData = Data("webkit-resume".utf8)
         let item = DownloadItem(
@@ -465,38 +278,4 @@ extension HarborModelAndSafetyTests {
         XCTAssertNil(decoded.resumeData)
     }
 
-    func testPayloadResolutionRejectsRootAndTraversalAndKeepsExactPayloadPaths() throws {
-        let fileManager = FileManager.default
-        let rootURL = fileManager.temporaryDirectory
-            .appendingPathComponent(UUID().uuidString, isDirectory: true)
-        let destinationURL = rootURL.appendingPathComponent("destination", isDirectory: true)
-        let outsideURL = rootURL.appendingPathComponent("outside.bin")
-        try fileManager.createDirectory(at: destinationURL, withIntermediateDirectories: true)
-        defer { try? fileManager.removeItem(at: rootURL) }
-
-        let resolution = DownloadDataRemovalService().resolvePayloadURLs(
-            destinationFolderPath: destinationURL.path,
-            payloadPaths: [
-                "Collection/one.bin",
-                "Collection/two.bin",
-                destinationURL.appendingPathComponent("single.iso").path,
-                destinationURL.path,
-                "../outside.bin",
-                outsideURL.path
-            ]
-        )
-
-        XCTAssertEqual(
-            resolution.safeURLs.map(\.path),
-            [
-                destinationURL.appendingPathComponent("Collection/one.bin").path,
-                destinationURL.appendingPathComponent("Collection/two.bin").path,
-                destinationURL.appendingPathComponent("single.iso").path
-            ]
-        )
-        XCTAssertEqual(
-            Set(resolution.rejectedPaths),
-            Set([destinationURL.path, "../outside.bin", outsideURL.path])
-        )
-    }
 }
